# Pooled span dicitionary direct pointer