--- a/src/Paprika/Chain/PooledSpanDictionary.cs
+++ b/src/Paprika/Chain/PooledSpanDictionary.cs
@@ -13,6 +13,14 @@ namespace Paprika.Chain;
 /// </summary>
 public class PooledSpanDictionary : IDisposable
 {
+    /// <summary>
+    /// Gets the size of the address to the next item.
+    /// </summary>
+    /// <remarks>
+    /// Pointer size. Assumes 64 bits.     
+    /// </remarks>
+    private const int PointerSize = 8;
+
     private const int BufferSize = BufferPool.BufferSize;
 
     private readonly BufferPool _pool;
@@ -83,24 +91,21 @@ public class PooledSpanDictionary : IDisposable
     // How many bytes are used for preamble + hash leftover
     private const int PreambleLength = 3;
 
-    private const int AddressLength = 4;
+    private const int AddressLength = PointerSize;
 
     private const int KeyLengthLength = 1;
     private const int ValueLengthLength = 2;
 
-    private unsafe SearchResult TryGetImpl(scoped ReadOnlySpan<byte> key, uint leftover, uint bucket)
+    private SearchResult TryGetImpl(scoped ReadOnlySpan<byte> key, uint leftover, uint bucket)
     {
         Debug.Assert(BitOperations.LeadingZeroCount(leftover) >= 11, "First 10 bits should be left unused");
 
         var address = _root[(int)bucket];
-        if (address == 0) goto NotFound;
+        if (address == UIntPtr.Zero) goto NotFound;
 
-        ref var pages = ref MemoryMarshal.GetReference(CollectionsMarshal.AsSpan(_pages));
         do
         {
-            var (pageNo, atPage) = Math.DivRem(address, Page.PageSize);
-
-            ref var at = ref Unsafe.AsRef<byte>((byte*)Unsafe.Add(ref pages, pageNo).Raw.ToPointer() + atPage);
+            ref var at = ref ReadAtAddress(address);
 
             var header = at & PreambleBits;
             if ((header & DestroyedBit) == 0)
@@ -126,11 +131,13 @@ public class PooledSpanDictionary : IDisposable
             }
 
             // Decode next entry address
-            address = Unsafe.ReadUnaligned<uint>(ref Unsafe.Add(ref at, PreambleLength));
+            address = Unsafe.ReadUnaligned<UIntPtr>(ref Unsafe.Add(ref at, PreambleLength));
         } while (address != 0);
     NotFound:
         return default;
     }
+
+    private static unsafe ref byte ReadAtAddress(UIntPtr address) => ref Unsafe.AsRef<byte>(address.ToPointer());
 
     private static (uint leftover, uint bucket) GetBucketAndLeftover(ulong hash)
         => Math.DivRem(Mix(hash), Root.BucketCount);
@@ -267,7 +274,7 @@ public class PooledSpanDictionary : IDisposable
 
         Debug.Assert(BitOperations.LeadingZeroCount(leftover) >= 10, "First 10 bits should be left unused");
 
-        var root = _root[(int)bucket];
+        UIntPtr root = _root[(int)bucket];
 
         var dataLength = data1.Length + data0.Length;
 
@@ -315,10 +322,10 @@ public class PooledSpanDictionary : IDisposable
     /// Enumerator walks through all the values beside the ones that were destroyed in this dictionary
     /// with <see cref="PooledSpanDictionary.Destroy"/>.
     /// </summary>
-    public ref struct Enumerator(PooledSpanDictionary dictionary)
+    public unsafe ref struct Enumerator(PooledSpanDictionary dictionary)
     {
         private int _bucket = -1;
-        private uint _address = 0;
+        private UIntPtr _address = 0;
         private ref byte _at;
 
         public bool MoveNext()
@@ -326,7 +333,7 @@ public class PooledSpanDictionary : IDisposable
             while (_bucket < Root.BucketCount)
             {
                 // On empty, scan to the next bucket that is not empty
-                while (_address == 0)
+                while (_address == UIntPtr.Zero)
                 {
                     _bucket++;
                     if (_bucket == Root.BucketCount)
@@ -338,13 +345,13 @@ public class PooledSpanDictionary : IDisposable
                 }
 
                 // Scan the bucket till it's not destroyed
-                while (_address != 0)
+                while (_address != UIntPtr.Zero)
                 {
                     // Capture the current, move address to next immediately
-                    ref var at = ref dictionary.GetAt(_address);
+                    ref var at = ref ReadAtAddress(_address);
 
                     // The position is captured in ref at above, move to next
-                    _address = Unsafe.ReadUnaligned<uint>(ref Unsafe.Add(ref at, PreambleLength));
+                    _address = Unsafe.ReadUnaligned<UIntPtr>(ref Unsafe.Add(ref at, PreambleLength));
 
                     if ((at & DestroyedBit) == 0)
                     {
@@ -414,14 +421,6 @@ public class PooledSpanDictionary : IDisposable
         }
     }
 
-    private ref byte GetAt(uint address)
-    {
-        Debug.Assert(address > 0);
-
-        var (pageNo, atPage) = Math.DivRem(address, Page.PageSize);
-        return ref Unsafe.Add(ref MemoryMarshal.GetReference(_pages[(int)pageNo].Span), (int)atPage);
-    }
-
     private void AllocateNewPage()
     {
         var page = RentNewPage(false);
@@ -440,7 +439,7 @@ public class PooledSpanDictionary : IDisposable
     private static uint Mix(ulong hash) => unchecked((uint)((hash >> 32) ^ hash));
 
 
-    private Span<byte> Write(int size, out uint addr)
+    private Span<byte> Write(int size, out UIntPtr addr)
     {
         if (BufferSize - _position < size)
         {
@@ -450,10 +449,9 @@ public class PooledSpanDictionary : IDisposable
 
         // allocated before the position is changed
         var span = _current.Span.Slice(_position, size);
+        addr = _current.Raw + (UIntPtr)_position;
 
-        addr = (uint)(_position + (_pages.Count - 1) * BufferSize);
         _position += size;
-
         return span;
     }
 
@@ -509,27 +507,44 @@ public class PooledSpanDictionary : IDisposable
         static string S(in NibblePath full) => full.UnsafeAsKeccak.ToString();
     }
 
-    private readonly struct Root(Page[] pages)
+    [StructLayout(LayoutKind.Explicit, Size = SizeOf)]
+    private readonly struct Root
     {
+        [FieldOffset(0)]
+        private readonly Page _pages;
+
+        public Root(Page[] pages)
+        {
+            Debug.Assert(pages.Length == PageCount);
+            pages.CopyTo(MemoryMarshal.CreateSpan(ref _pages, PageCount));
+        }
+
         /// <summary>
-        /// 16gives 4kb * 16, 64kb allocated per dictionary.
-        /// This gives 16k buckets which should be sufficient to have a really low ratio of collisions for majority of the blocks.
+        /// The size of this structure.
+        /// </summary>
+        private const int SizeOf = PageCount * PointerSize;
+
+        /// <summary>
+        /// 16 pages, gives 4kb * 16, 64kb of memory allocated per dictionary.
+        /// This gives 8k buckets (<see cref="PooledSpanDictionary.PointerSize"/>
+        /// which should be sufficient to have a really low ratio of collisions for the majority of the blocks.
         /// </summary>
         public const int PageCount = 16;
 
         public static readonly int BucketCountLog2 = BitOperations.Log2(BucketCount);
 
         public const int BucketCount = PageCount * BucketsPerPage;
-        private const int BucketsPerPage = Page.PageSize / sizeof(uint);
+        private const int BucketsPerPage = Page.PageSize / PointerSize;
         private const int InPageMask = BucketsPerPage - 1;
         private static readonly int PageShift = BitOperations.Log2(BucketsPerPage);
 
-        public unsafe ref uint this[int bucket]
+        public unsafe ref UIntPtr this[int bucket]
         {
             get
             {
-                var raw = Unsafe.Add(ref MemoryMarshal.GetArrayDataReference(pages), bucket >> PageShift).Raw;
-                return ref Unsafe.Add(ref Unsafe.AsRef<uint>(raw.ToPointer()), bucket & InPageMask);
+                var shift = bucket >> PageShift;
+                var raw = Unsafe.Add(ref Unsafe.AsRef(in _pages), shift).Raw;
+                return ref Unsafe.Add(ref Unsafe.AsRef<UIntPtr>(raw.ToPointer()), bucket & InPageMask);
             }
         }
     }